Fix room camera switch by using two area2d for detection
The problem was that you could still be inside the first area with part of the player's body. The camera switched to the new room, but if you were to move back fully into the first area you wouldn't get an area entered triggered because you were still in it from before.